nixos/networkd: improved behavior with networking.enableIPv6=false
Minor refactoring to set correct options when ipv6 is disabled,
namely:
- Configure DHCP=ipv4
- deactivate linklocal addresses configuration and router advertisment
  listen

In the process, a bug was fixed in the DHCP option declaration which
excpected "v4" or "v6" whereas upstream documentation says it should
be "ipv6" or "ipv4".

Also, improved the system by not generating 99-main if no global options
(networking.useDHCP, defaultGateways, search/domains) are configured,
this prevent breaking interfaces such as tunnels when it is not needed
(otherwise interfaces end up in the "configuring" state, see #30904).

Conflicts:
	nixos/modules/system/boot/networkd.nix

Conflicts:
	nixos/modules/system/boot/networkd.nix
	nixos/modules/tasks/network-interfaces-systemd.nix